+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [5.5.0](https://github.com/plivo/plivo-java/tree/v5.5.0) (2021-11-15)
 **Features - Voice: Multiparty calls**
+- The [Add Multiparty Call API](https://www.plivo.com/docs/voice/api/multiparty-call/participants#add-a-participant) allows for greater functionality by accepting options like `start recording audio`, `stop recording audio`, and their HTTP methods.
 - [Multiparty Calls](https://www.plivo.com/docs/voice/api/multiparty-call/) now has new APIs to `stop` and `play` audio.
 
 ## [5.4.1](https://github.com/plivo/plivo-java/tree/v5.4.1) (2021-11-05)